@@ -2,8 +2,8 @@
     "name": "rmcgirr83/activationjustification",
     "type": "phpbb-extension",
     "description": "How do you know whether or not the person who has just signed up for your board signed up just to spam it? Letting people justify their activation can help you decide exactly who should be activated and who shouldn't be. The extension is only active if user registration in the ACP is set to `By Admin`.",
-    "version": "1.0.2",
-    "time": "2016-01-16",
+    "version": "1.0.3",
+    "time": "2021-01-06",
     "keywords": [
         "phpbb",
         "extension",
@@ -20,7 +20,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3"
+        "php": ">=7.1"
     },
     "require-dev": {
         "phpbb/epv": "dev-master"
@@ -28,7 +28,7 @@
     "extra": {
         "display-name": "Activation Justification",
         "soft-require": {
-            "phpbb/phpbb": ">=3.1.4-RC1,<3.3@dev"
+            "phpbb/phpbb": ">=3.2.11"
         },
         "version-check": {
             "host": "www.phpbb.com",
