@@ -24,6 +24,6 @@ class ext extends \phpbb\extension\base
 	public function is_enableable()
 	{
 		$config = $this->container->get('config');
-		return version_compare($config['version'], '3.1.4-RC1', '>=');
+		return version_compare($config['version'], '3.2.11', '>=');
 	}
 }
